style: refactor component structure
